% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clincluster}
\alias{clincluster}
\title{integration of different datasets}
\usage{
clincluster(object, metadata, groups)
}
\arguments{
\item{object}{seurat object from Prepare_Seurat}

\item{metadata}{metadata from Prepare_Metadata}

\item{groups}{how many groups you expect to be in your final clusters}
}
\value{

}
\description{
integration of different datasets
}
