% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Prepare_Metadata}
\alias{Prepare_Metadata}
\title{Prepare metadata of the seurat object}
\usage{
Prepare_Metadata(object)
}
\arguments{
\item{object}{seurat object from Prepare_seurat}
}
\value{

}
\description{
Prepare metadata of the seurat object
}
