% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Prepare_Seurat}
\alias{Prepare_Seurat}
\title{Prepare the seurat object consisting of different datasets}
\usage{
Prepare_Seurat(seurat_list, sample_names)
}
\arguments{
\item{seurat_list}{a list of seurat objects}

\item{sample_names}{a vector of names for seurat objects or samples}
}
\value{
Returns a Seurat object that can be used for following analysis
}
\description{
Prepare the seurat object consisting of different datasets
}
